Refactored type_editor module into 5 different modules inside a editors/ subfolder
