+++ libasvat/imgui/fonts.py
@@ -1,9 +1,9 @@
 import os
 import libasvat.command_utils as cmd_utils
-import libasvat.imgui.type_editor as types
 from libasvat.imgui.math import Vector2
 from libasvat.imgui.general import drop_down
 from libasvat.imgui.colors import Colors
+from libasvat.imgui.editors import TypeDatabase, TypeEditor
 from libasvat.utils import get_all_files
 from contextlib import contextmanager
 from imgui_bundle import imgui
@@ -24,8 +24,8 @@ class FontID(str):
     """
 
 
-@types.TypeDatabase.register_editor_for_type(FontID)
-class FontIDEditor(types.TypeEditor):
+@TypeDatabase.register_editor_for_type(FontID)
+class FontIDEditor(TypeEditor):
     """Imgui TypeEditor for selecting a FontID value."""
 
     def __init__(self, config: dict):
